refactor: Merge buffer_tools.py and process_tools.py into sandbox_common.py
Fold buffer tool builders and process/workspace tool builders into
sandbox_common.py. Delete both standalone files. Update all import
references in sandbox_storage_tools.py and tests.

Net: -2 files, tools directory 14 → 12 files.
